refactor(frontend): move ChainSelect
